+++ graphite_influxdb/__init__.py
@@ -1,1 +1,1 @@
-from graphite_influxdb import *
+from classes import *
